move reaction damage calc to combat
