+++ pkg/reactable/overload.go
@@ -57,7 +57,7 @@ func (r *Reactable) TryOverload(a *info.AttackEvent) bool {
 		}
 		char := r.core.Player.ByIndex(a.Info.ActorIndex)
 		em := char.Stat(attributes.EM)
-		flatdmg, snap := calcReactionDmg(char, atk, em)
+		flatdmg, snap := combat.CalcReactionDmg(char.Base.Level, char, atk, em)
 		atk.FlatDmg = 2.75 * flatdmg
 		r.core.QueueAttackWithSnap(atk, snap, combat.NewCircleHitOnTarget(r.self, nil, 3), 1)
 	}
