+++ pkg/reactable/reactable.go
@@ -8,9 +8,9 @@ import (
 
 	"github.com/genshinsim/gcsim/pkg/core"
 	"github.com/genshinsim/gcsim/pkg/core/attributes"
+	"github.com/genshinsim/gcsim/pkg/core/combat"
 	"github.com/genshinsim/gcsim/pkg/core/event"
 	"github.com/genshinsim/gcsim/pkg/core/info"
-	"github.com/genshinsim/gcsim/pkg/core/player/character"
 )
 
 type Modifier int
@@ -461,122 +461,7 @@ func (r *Reactable) Tick() {
 	}
 }
 
-func calcReactionDmg(char *character.CharWrapper, atk info.AttackInfo, em float64) (float64, info.Snapshot) {
-	lvl := char.Base.Level - 1
-	if lvl > 89 {
-		lvl = 89
-	}
-	if lvl < 0 {
-		lvl = 0
-	}
-	snap := info.Snapshot{
-		CharLvl: char.Base.Level,
-	}
-	snap.Stats[attributes.EM] = em
-	return (1 + ((16 * em) / (2000 + em)) + char.ReactBonus(atk)) * reactionLvlBase[lvl], snap
-}
-
 func (r *Reactable) calcCatalyzeDmg(atk info.AttackInfo, em float64) float64 {
 	char := r.core.Player.ByIndex(atk.ActorIndex)
-	lvl := char.Base.Level - 1
-	if lvl > 89 {
-		lvl = 89
-	}
-	if lvl < 0 {
-		lvl = 0
-	}
-	return (1 + ((5 * em) / (1200 + em)) + r.core.Player.ByIndex(atk.ActorIndex).ReactBonus(atk)) * reactionLvlBase[lvl]
-}
-
-var reactionLvlBase = []float64{
-	17.1656055450439,
-	18.5350475311279,
-	19.9048538208007,
-	21.27490234375,
-	22.6453990936279,
-	24.6496124267578,
-	26.6406421661376,
-	28.8685874938964,
-	31.3676795959472,
-	34.1433448791503,
-	37.201000213623,
-	40.6599998474121,
-	44.4466667175292,
-	48.5635185241699,
-	53.7484817504882,
-	59.0818977355957,
-	64.4200439453125,
-	69.7244567871093,
-	75.1231384277343,
-	80.5847778320312,
-	86.1120300292968,
-	91.703742980957,
-	97.24462890625,
-	102.812644958496,
-	108.409561157226,
-	113.201690673828,
-	118.102905273437,
-	122.979316711425,
-	129.727325439453,
-	136.292907714843,
-	142.670852661132,
-	149.029022216796,
-	155.4169921875,
-	161.825500488281,
-	169.106307983398,
-	176.518081665039,
-	184.07273864746,
-	191.709518432617,
-	199.556915283203,
-	207.382049560546,
-	215.398895263671,
-	224.165664672851,
-	233.502166748046,
-	243.35057067871,
-	256.063079833984,
-	268.543487548828,
-	281.526062011718,
-	295.013641357421,
-	309.067199707031,
-	323.601593017578,
-	336.757537841796,
-	350.530303955078,
-	364.482696533203,
-	378.619171142578,
-	398.600402832031,
-	416.398254394531,
-	434.386993408203,
-	452.951049804687,
-	472.606231689453,
-	492.884887695312,
-	513.568542480468,
-	539.103210449218,
-	565.510559082031,
-	592.538757324218,
-	624.443420410156,
-	651.470153808593,
-	679.496826171875,
-	707.794067382812,
-	736.671447753906,
-	765.640258789062,
-	794.773376464843,
-	824.677368164062,
-	851.157775878906,
-	877.742065429687,
-	914.229125976562,
-	946.746765136718,
-	979.411376953125,
-	1011.22302246093,
-	1044.79174804687,
-	1077.44372558593,
-	1109.99755859375,
-	1142.9765625,
-	1176.36950683593,
-	1210.18444824218,
-	1253.83569335937,
-	1288.95275878906,
-	1325.48413085937,
-	1363.45690917968,
-	1405.09741210937,
-	1446.853515625,
+	return combat.CalcCatalyzeDmg(char.Base.Level, char, atk, em)
 }
